requirements.txt
accelerate>=0.26.0
av>=16.1.0
diffusers>=0.31.0
matplotlib>=3.5.0
numpy>=1.26.4
opencv-python>=4.8.0
pandas>=2.0.0
tokenizers>=0.21.0
torcheval>=0.0.7
tqdm>=4.60.0
transformers>=4.51.0
pillow>=10.0.0